2.0.0 release refactor for requests/httpx separation FaultTolerantAuthRequestSession is not needed anymore, error for refresh is handled with a session method override instead.
