+++ annotell-auth/annotell/auth/__init__.py
@@ -3,6 +3,6 @@ from logging import NullHandler
 
 logging.getLogger(__name__).addHandler(NullHandler())
 
-__version__ = "1.8.1"
+__version__ = "2.0.0"
 
 DEFAULT_HOST = "https://user.annotell.com"
